♻️ Pass stepLogger and openOptions intead of Options from Open to wrap.Open
