+++ internal/wrap/open.go
@@ -13,13 +13,8 @@ type OpenOptions = opts.OpenOptions
 
 var DefaultOpenOptions = opts.DefaultOpenOptions
 
-func Open(ctx context.Context, driverName, dsn string, opts ...Option) (*sql.DB, error) {
-	options := NewOptions(driverName, opts...)
-	logger := NewStepLogger(options.Logger, DurationAttrFunc(options.DurationKey, options.DurationType))
-
-	openOptions := buildOpenOptions(options)
-
-	lg := logger.With(
+func Open(ctx context.Context, driverName, dsn string, stepLogger *StepLogger, openOptions *OpenOptions) (*sql.DB, error) {
+	lg := stepLogger.With(
 		slog.String("driver", driverName),
 		slog.String("dsn", dsn),
 	)
@@ -27,7 +22,7 @@ func Open(ctx context.Context, driverName, dsn string, opts ...Option) (*sql.DB,
 	var db *sql.DB
 	err := IgnoreAttr(lg.Step(ctx, openOptions.Open, func() (*slog.Attr, error) {
 		var err error
-		db, err = open(driverName, dsn, logger, openOptions)
+		db, err = open(driverName, dsn, stepLogger, openOptions)
 		return nil, err
 	}))
 	if err != nil {
@@ -58,56 +53,4 @@ func openWithDriver(drv driver.Driver, dsn string, logger *StepLogger, connector
 		origConnector = &dsnConnector{dsn: dsn, driver: drv}
 	}
 	return sql.OpenDB(WrapConnector(origConnector, logger, connectorOptions)), nil
-}
-
-func buildOpenOptions(options *Options) *OpenOptions {
-	connOptions := &ConnOptions{
-		IDGen:   options.IDGen,
-		Begin:   &options.ConnBegin,
-		BeginTx: &options.ConnBeginTx,
-		TxIDKey: options.TxIDKey,
-		Tx: &TxOptions{
-			Commit:   &options.TxCommit,
-			Rollback: &options.TxRollback,
-		},
-		Close:          &options.ConnClose,
-		Prepare:        &options.ConnPrepare,
-		PrepareContext: &options.ConnPrepareContext,
-		StmtIDKey:      options.StmtIDKey,
-		Stmt: &StmtOptions{
-			Close:        &options.StmtClose,
-			Exec:         &options.StmtExec,
-			Query:        &options.StmtQuery,
-			ExecContext:  &options.StmtExecContext,
-			QueryContext: &options.StmtQueryContext,
-			Rows: &RowsOptions{
-				Close:         &options.RowsClose,
-				Next:          &options.RowsNext,
-				NextResultSet: &options.RowsNextResultSet,
-			},
-		},
-		ResetSession: &options.ConnResetSession,
-		Ping:         &options.ConnPing,
-		ExecContext:  &options.ConnExecContext,
-		QueryContext: &options.ConnQueryContext,
-		Rows: &RowsOptions{
-			Close:         &options.RowsClose,
-			Next:          &options.RowsNext,
-			NextResultSet: &options.RowsNextResultSet,
-		},
-	}
-	return &OpenOptions{
-		Open: &options.SqlslogOpen,
-		Driver: &DriverOptions{
-			IDGen:         options.IDGen,
-			ConnIDKey:     options.ConnIDKey,
-			Open:          &options.DriverOpen,
-			OpenConnector: &options.DriverOpenConnector,
-			Conn:          connOptions,
-			Connector: &ConnectorOptions{
-				Connect: &options.ConnectorConnect,
-				Conn:    connOptions,
-			},
-		},
-	}
 }
